requirements.txt
streamlit==1.39.0
requests==2.31.0
numpy==2.3.2
pandas==2.2.3
plotly==5.17.0
python-dotenv==1.0.0
pytz==2023.3
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3
streamlit-card==0.0.61
streamlit-extras==0.3.5
